perf(cp): drop explicit .float() casts in arbitrary fwd/bwd accumulators
cw-dfw round-7 (4720849f, content-keyed func cache) reached 76.8 ms /
step / 85% throughput ratio. The remaining gap to DP4 is dominated
by HBM churn from `.float()` casts: every `out_local += partial.float()`
materialises a transient ~112 MB fp32 tensor before adding into the
accumulator. Across 8 layers × cp_size=2 ring steps that's ~1.8 GB
of throwaway fp32 allocs in forward, and ~5.6 GB in backward (3
gradients × cp_size ring steps × 8 layers, plus the reverse-ring
dKV exchange).

Fix: just remove the explicit `.float()` calls. PyTorch's `add_`
handles mixed-dtype promotion in-register inside the add kernel —
no HBM temp, no extra copy. Math is identical to the cast-then-add
path (both do bf16 → fp32 → add to fp32 accumulator); only the
allocation is gone. Final `.to(q_local.dtype)` cast back to bf16 at
return is unchanged.

Scope: only the `_arbitrary` paths (the only ones the wrapper
actually routes through). The legacy `_multi_gpu_forward` /
`_multi_gpu_backward` 4-op paths still have their `.float()` calls
untouched — they're dead code in v0 but kept for tests that
deliberately exercise them.

Predicted cw-dfw round-8 effect: HBM bandwidth + allocator pressure
drop → save ~5–8 ms / step → 76.8 → ~70 ms → ratio crosses 90%.
Loss curve unchanged (numerically identical math).

+++ examples/hstu/context_parallel/hstu_attn_cp.py
@@ -1141,7 +1141,13 @@ def _multi_gpu_forward_arbitrary(
             func=func,
             quant_mode=-1,
         )
-        out_local += partial.float()
+        # In-place add with mixed dtype: PyTorch promotes `partial` (bf16)
+        # to fp32 inside the add kernel — no transient HBM allocation,
+        # unlike the explicit `.float()` cast which materialises a fp32
+        # temp of size ~112 MB per ring step. Across 8 layers × cp_size
+        # ring steps this saves ~1.8 GB of HBM churn per training step
+        # (cw-dfw round-7 → round-8 optimisation).
+        out_local += partial
 
         # 3. Wait for P2P + swap.
         if step < cp_size - 1:
@@ -1673,11 +1679,18 @@ def _multi_gpu_backward_arbitrary(
             window_size=(-1, -1),
             func=func,
         )
-        dq_acc += dq_p.float()
+        # Mixed-dtype in-place add: avoids the explicit `.float()` cast
+        # which would allocate a transient ~112 MB fp32 temp per
+        # gradient per ring step. PyTorch's add_ kernel handles
+        # bf16 → fp32 promotion in-register. Saves ~5–6 GB of HBM
+        # churn per training step on cw-dfw (8 layers × cp_size × 3
+        # gradients). Same cw-dfw round-7 → round-8 optimisation as
+        # the forward accumulator.
+        dq_acc += dq_p
         if step == 0:
             # peer == self; dK/dV are ours.
-            dk_acc += dk_p.float()
-            dv_acc += dv_p.float()
+            dk_acc += dk_p
+            dv_acc += dv_p
         else:
             dkv_to_send.append((step, dk_p, dv_p))
 
@@ -1713,8 +1726,10 @@ def _multi_gpu_backward_arbitrary(
         )
         for r in reqs_v:
             r.wait()
-        dk_acc += recv_dk.float()
-        dv_acc += recv_dv.float()
+        # Mixed-dtype in-place add: see fwd accumulator comment for
+        # why the explicit `.float()` cast was wasteful.
+        dk_acc += recv_dk
+        dv_acc += recv_dv
 
     return (
         dq_acc.to(q_local.dtype),
